Cannot add conflicting class
I changed some things around in order to check for conflicting classes. However, without the pop up UI for conflicts, it appears to the user that you cannot check a box that is a conflict. Eventually it will be able to be checked then the pop up will appear than the corresponding action will occur.